+++ src/aui-scheduler/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1309](https://issues.liferay.com/browse/AUI-1309) Make aui-scheduler responsive
 * [AUI-1809](https://issues.liferay.com/browse/AUI-1809) Events with duration longer than a week don't render correctly on month view
 * [AUI-1640](https://issues.liferay.com/browse/AUI-1640) Scheduler should create SchedulerEvent instances lazily
 * [AUI-1771](https://issues.liferay.com/browse/AUI-1771) Display all events in scheduler's month view